Local time
formateDateTime() function receives a parameter named timezone which passes through the timezone data (response.data.timezone) fetched from the openweathermap.org API.
As the timezone data represents a shift in seconds from UTC, it needs to be divided by 3600 in order to convert the timezone seconds into hours. Also, instead of the getHours() method, getUTCHours() was used to calculate the local time of any given location in respect of UTC time and not user's local time.
When getHours() is used, the difference between user's local and UTC time needs to be added or subtracted in the calculation because this method returns the hours in the specified date according to local time. Obviously, this is not a good practice as the difference between one user's local time and the UTC is not the same for any other user.