Weather icon
Created an object weatherIcon to store the weather condition codes group names from the owm API and their compatible erikflowers weather icon names.
Created a function getWeatherIcon that generates a compatible weather icon based on the API's weather id. The function takes 2 parameters - weatherIcon and rangeId. When called, the weatherIcon object and the retrieved API weather id (response.data.weather[0].id) are passed as parameters. Multiple if..else statements are executed until the range that the weather id value belongs to, is determined. Next, based on the range the weatherIcon property/key, which value is the erikflowers icon name, is determined.